Successfuly deployed arithmetic functions and UI
Thus far, the calculator successfuly executes addition, subtraction,
multiplication and division on a single pair of number at a time. It
doesn't allow the user to input operators if the calculator display is
blank and the user is unable to enter two consecutive operators.

Flexboxes are used to achieve the keypad layout.